Add a function to the HydraIface which creates a complexmodel object out of an Iface object
git-svn-id: svn+ssh://ec2-54-229-95-247.eu-west-1.compute.amazonaws.com/home/svn/svnrepository/HYDRA@72 5b2e1b5e-a1f6-4386-8dd1-2f88c1042e8b

+++ HydraServer/trunk/python/soap_server/hydra_complexmodels.py
@@ -1,5 +1,5 @@
 from spyne.model.complex import Array, ComplexModel
-from spyne.model.primitive import String, Integer, Double
+from spyne.model.primitive import String, Integer, Decimal
 
 class Attribute(ComplexModel):
  _type_info = [
@@ -20,8 +20,8 @@ class Node(ComplexModel):
         ('node_id', Integer),
         ('node_name', String),
         ('node_description', String),
-        ('node_x', Double),
-        ('node_y', Double),
+        ('node_x', Decimal),
+        ('node_y', Decimal),
         ('attributes', Array(Attribute)),
     ]
